pyproject.toml
[project]
name = "chemputation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipython>=9.9.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pango>=0.0.1",
    "pillow>=12.1.0",
    "pycairo>=1.29.0",
    "rdkit>=2025.9.3",
    "seaborn>=0.13.2",
    "weasyprint>=67.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rxnframe"]
